requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic
pytest==8.0.0
httpx==0.26.0
psycopg2-binary